Rewrite README: simplify and add graphic, links
Graphics and links to The Chromium Projects website help the user
understand what chromeos-resize does, how, and why.

All recommendations are moved to the "Considerations" subsection of
"Usage", and the main body text now exclusively describes operation.

Using the same names for partitions as The Chromium Projects site do
remove ambiguity and jargon to make the text clearer.

Finally, from personal experience, I've learned that 3 GiB is quite
small for the user state partition, after installing the Google Play
Store and being unable to remove it.

Signed-off-by: Ethan Madison <ethan@ethanmad.com>